+++ temporian/beam/operators/window/moving_max.py
@@ -15,21 +15,23 @@
 from typing import Type
 from temporian.beam.operators.window.base import BaseWindowBeamImplementation
 
-from temporian.core.operators.window.moving_sum import MovingSumOperator
+from temporian.core.operators.window.moving_max import (
+    MovingMaxOperator,
+)
 from temporian.beam import implementation_lib
 from temporian.implementation.numpy.operators.window.base import (
     BaseWindowNumpyImplementation,
 )
-from temporian.implementation.numpy.operators.window.moving_sum import (
-    MovingSumNumpyImplementation,
+from temporian.implementation.numpy.operators.window.moving_max import (
+    MovingMaxNumpyImplementation,
 )
 
 
-class MovingSumBeamImplementation(BaseWindowBeamImplementation):
+class MovingMaxBeamImplementation(BaseWindowBeamImplementation):
     def _implementation(self) -> Type[BaseWindowNumpyImplementation]:
-        return MovingSumNumpyImplementation
+        return MovingMaxNumpyImplementation
 
 
 implementation_lib.register_operator_implementation(
-    MovingSumOperator, MovingSumBeamImplementation
+    MovingMaxOperator, MovingMaxBeamImplementation
 )