+++ temporian/implementation/numpy/test/registered_operators_test.py
@@ -95,6 +95,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "TICK",
             "TIMESTAMPS",
             "UNIQUE_TIMESTAMPS",
+            "UNTIL_NEXT",
             "XOR",
         ]