+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added `steps` argument to `EventSet.since_last()` operator.
 - Added variable `window_length` option to moving window operators.
 - Added unsupervised anomaly detection tutorial.
+- Add `until_next` operator.
 - Added Beam execution tutorial.
 - Added changelog to docs site.
 
@@ -31,7 +32,6 @@
 ### Fixes
 
 - Fixed tutorials opening unreleased versions of the notebooks.
--
 
 ## 0.1.3